services/ocr/requirements.txt
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
numpy==1.26.2
Pillow==10.1.0
